requirements.txt
streamlit
pytesseract
pdfplumber
Pillow
python-docx
